# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rasheed - Finance Savings Agent",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "قائمة التصنيفات",
        "description": "يرجّع التصنيفات الافتراضية بالإضافة للتصنيفات الخاصة بالمستخدم. `category_type` يحدد إذا كان التصنيف للمصروف أو الدخل أو كليهما (`both`).",
        "operationId": "list_categories_api_v1_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryOut"
                  },
                  "type": "array",
                  "title": "Response List Categories Api V1 Categories Get"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "إنشاء تصنيف خاص",
        "description": "ينشئ تصنيفًا خاصًا بالمستخدم بالإضافة للتصنيفات الافتراضية.",
        "operationId": "create_category_api_v1_categories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOut"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/goals": {
      "get": {
        "tags": [
          "Goals"
        ],
        "summary": "قائمة الأهداف",
        "description": "يرجّع كل أهداف المستخدم مرتبة حسب الحالة ثم الأولوية. يُطبَّق عند القراءة التصفير الشهري للأهداف المتكررة. `progress_percentage` محسوبة من `current_amount` و`target_amount`.",
        "operationId": "list_goals_api_v1_goals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GoalOut"
                  },
                  "type": "array",
                  "title": "Response List Goals Api V1 Goals Get"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Goals"
        ],
        "summary": "إنشاء هدف",
        "description": "ينشئ هدفًا جديدًا بأولوية في آخر الترتيب. إذا كان `is_recurring=true` يصبح «مصروفًا شهريًا ثابتًا»: يُصفَّر تلقائيًا في بداية كل شهر بعد أن يكتمل في الشهر السابق.",
        "operationId": "create_goal_api_v1_goals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalOut"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/goals/reorder": {
      "put": {
        "tags": [
          "Goals"
        ],
        "summary": "إعادة ترتيب الأولويات",
        "description": "يحدّد الترتيب الجديد لكل الأهداف النشطة (من الأعلى أولوية للأقل). يجب إرسال معرّفات كل الأهداف النشطة الحالية بالضبط، لا أكثر ولا أقل، لتفادي ترتيب جزئي.",
        "operationId": "reorder_goals_api_v1_goals_reorder_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GoalOut"
                  },
                  "type": "array",
                  "title": "Response Reorder Goals Api V1 Goals Reorder Put"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "القائمة المرسلة لا تطابق مجموعة الأهداف النشطة الحالية.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/goals/{goal_id}/contribute": {
      "post": {
        "tags": [
          "Goals"
        ],
        "summary": "إضافة مبلغ لتقدّم هدف",
        "description": "يزيد `current_amount` للهدف بالمبلغ المرسل، ويحوّل حالته إلى `achieved` عند الوصول للمبلغ المستهدف. للأهداف النشطة فقط.",
        "operationId": "contribute_to_goal_api_v1_goals__goal_id__contribute_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalContribution"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalOut"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "الهدف غير نشط (منجز أو متروك).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "الهدف غير موجود.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{goal_id}": {
      "delete": {
        "tags": [
          "Goals"
        ],
        "summary": "حذف هدف",
        "description": "يحذف الهدف نهائيًا. لا يرجّع محتوى (204).",
        "operationId": "delete_goal_api_v1_goals__goal_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "الهدف غير موجود.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "اسم التصنيف"
          },
          "icon": {
            "type": "string",
            "title": "Icon",
            "description": "مفتاح أيقونة يفسّره تطبيق الموبايل",
            "default": "tag"
          },
          "category_type": {
            "$ref": "#/components/schemas/CategoryType",
            "description": "`expense` أو `income` أو `both`",
            "default": "expense"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CategoryCreate",
        "description": "تصنيف خاص ينشئه المستخدم بنفسه (بالإضافة للتصنيفات الافتراضية).",
        "examples": [
          {
            "category_type": "expense",
            "icon": "gift",
            "name": "هدايا"
          }
        ]
      },
      "CategoryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "description": "true = تصنيف افتراضي من النظام، false = أنشأه المستخدم"
          },
          "category_type": {
            "$ref": "#/components/schemas/CategoryType"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "icon",
          "is_default",
          "category_type"
        ],
        "title": "CategoryOut",
        "examples": [
          {
            "category_type": "expense",
            "icon": "restaurant",
            "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "is_default": true,
            "name": "مطاعم وكافيهات"
          }
        ]
      },
      "CategoryType": {
        "type": "string",
        "enum": [
          "expense",
          "income",
          "both"
        ],
        "title": "CategoryType",
        "description": "نوع التصنيف: بيحدد وين يظهر بواجهة التسجيل السريع.\nEXPENSE تظهر لما المستخدم يسجّل مصروف، INCOME لما يسجّل دخل،\nBOTH تظهر بالحالتين (مثال: \"تحويل بين حسابات\")."
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "رسالة الخطأ (بالعربي)"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "شكل الخطأ القياسي الراجع من كل الـ Endpoints.",
        "examples": [
          {
            "detail": "الهدف غير موجود"
          }
        ]
      },
      "GoalContribution": {
        "properties": {
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "description": "المبلغ المضاف بالدينار"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "GoalContribution",
        "description": "إضافة مبلغ لتقدم الهدف (مثلاً بعد ما المستخدم يوفر مبلغ فعليًا).",
        "examples": [
          {
            "amount": 50
          }
        ]
      },
      "GoalCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 150,
            "minLength": 2,
            "title": "Title",
            "description": "عنوان الهدف"
          },
          "icon": {
            "type": "string",
            "title": "Icon",
            "description": "مفتاح أيقونة يفسّره تطبيق الموبايل (مثل laptop أو plane)",
            "default": "target"
          },
          "target_amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Target Amount",
            "description": "المبلغ المستهدف بالدينار"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline",
            "description": "الموعد النهائي (اختياري)"
          },
          "is_recurring": {
            "type": "boolean",
            "title": "Is Recurring",
            "description": "true = مصروف شهري ثابت يُصفَّر تلقائيًا كل شهر بعد اكتماله",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "title",
          "target_amount"
        ],
        "title": "GoalCreate",
        "examples": [
          {
            "deadline": "2027-01-05",
            "icon": "laptop",
            "is_recurring": false,
            "target_amount": 600,
            "title": "لابتوب جديد"
          },
          {
            "icon": "dumbbell",
            "is_recurring": true,
            "target_amount": 40,
            "title": "اشتراك النادي"
          }
        ]
      },
      "GoalOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "target_amount": {
            "type": "number",
            "title": "Target Amount"
          },
          "current_amount": {
            "type": "number",
            "title": "Current Amount"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "is_recurring": {
            "type": "boolean",
            "title": "Is Recurring"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "status": {
            "$ref": "#/components/schemas/GoalStatus"
          },
          "progress_percentage": {
            "type": "number",
            "title": "Progress Percentage",
            "description": "نسبة التقدم المئوية (0 إلى 100)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "icon",
          "target_amount",
          "current_amount",
          "priority",
          "is_recurring",
          "deadline",
          "status",
          "progress_percentage",
          "created_at"
        ],
        "title": "GoalOut",
        "examples": [
          {
            "created_at": "2026-08-08T10:47:45Z",
            "current_amount": 510,
            "deadline": "2027-01-05",
            "icon": "laptop",
            "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "is_recurring": false,
            "priority": 2,
            "progress_percentage": 85.0,
            "status": "active",
            "target_amount": 600,
            "title": "لابتوب جديد"
          }
        ]
      },
      "GoalReorderRequest": {
        "properties": {
          "ordered_goal_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Ordered Goal Ids",
            "description": "معرّفات الأهداف النشطة بالترتيب المطلوب (الأعلى أولوية أولًا)"
          }
        },
        "type": "object",
        "required": [
          "ordered_goal_ids"
        ],
        "title": "GoalReorderRequest",
        "description": "ترتيب جديد كامل لكل خطط المستخدم النشطة، كلستة IDs بالترتيب المطلوب.\nلازم تحتوي بالضبط نفس مجموعة الخطط النشطة الحالية — كل أو ولا شي،\nعشان نتجنب حالة نص خطط مرتبة ونص لأ.",
        "examples": [
          {
            "ordered_goal_ids": [
              "3fa85f64-5717-4562-b3fc-2c963f66afa6",
              "9c1d2e4a-7b3f-4d58-a1c6-0e5f8b2d7a91"
            ]
          }
        ]
      },
      "GoalStatus": {
        "type": "string",
        "enum": [
          "active",
          "achieved",
          "abandoned"
        ],
        "title": "GoalStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}